rename {_,}*py
this breaks existing functionality, but is a step in the right direction.
probably we should squash our work since the initial move into sidh/ into a
single commit to omit the existence of the _*py files at all.
