+++ sidh/csidh/poly_redc.py
@@ -1,8 +1,14 @@
-from sidh.csidh.poly_mul import *
+from sidh.csidh.poly_mul import poly_mul_middle, poly_mul_modxn
 
-# This function computes the reciprocal of a given polynomial
+fp = None
+def poly_redc_init(fp_):
+    global fp
+    fp = fp_
+
 def reciprocal(f, flen, n):
-
+    """
+    This function computes the reciprocal of a given polynomial
+    """
     if flen < n:
         # truncated reciprocal
         return reciprocal(f + [0] * (n - flen), n, n)
@@ -19,35 +25,35 @@ def reciprocal(f, flen, n):
     elif n == 2:
 
         # Base case when (1/f) = f_0 - f_1*x
-        return [f[0], fp_sub(0, f[1])], fp_sqr(f[0])
+        return [f[0],fp.fp_sub(0, f[1])],fp.fp_sqr(f[0])
 
     elif n == 3:
 
         g, a = reciprocal(f[:2], 2, 2)
 
-        t0 = fp_sqr(f[1])
-        t1 = fp_mul(f[0], f[2])
-        t2 = fp_sub(t1, t0)
-        t2 = fp_mul(t2, f[0])
-        return [fp_mul(g[0], a), fp_mul(g[1], a), fp_sub(0, t2)], fp_sqr(a)
+        t0 =fp.fp_sqr(f[1])
+        t1 =fp.fp_mul(f[0], f[2])
+        t2 =fp.fp_sub(t1, t0)
+        t2 =fp.fp_mul(t2, f[0])
+        return [fp.fp_mul(g[0], a),fp.fp_mul(g[1], a),fp.fp_sub(0, t2)],fp.fp_sqr(a)
 
     elif n == 4:
 
         # This case gives the same cost as the general case
         g, a = reciprocal(f[:2], 2, 2)
-        t0 = fp_sqr(f[1])
-        t1 = fp_mul(g[0], f[2])
-        t2 = fp_mul(g[0], f[3])
-        t3 = fp_mul(g[1], f[2])
-        t0 = fp_sub(t1, t0)
-        t1 = fp_add(t2, t3)
-        t2 = fp_mul(t0, g[0])
-        t3 = fp_mul(t0, g[1])
-        t4 = fp_mul(t1, g[0])
-        t3 = fp_add(t3, t4)
+        t0 =fp.fp_sqr(f[1])
+        t1 =fp.fp_mul(g[0], f[2])
+        t2 =fp.fp_mul(g[0], f[3])
+        t3 =fp.fp_mul(g[1], f[2])
+        t0 =fp.fp_sub(t1, t0)
+        t1 =fp.fp_add(t2, t3)
+        t2 =fp.fp_mul(t0, g[0])
+        t3 =fp.fp_mul(t0, g[1])
+        t4 =fp.fp_mul(t1, g[0])
+        t3 =fp.fp_add(t3, t4)
         return (
-            [fp_mul(g[0], a), fp_mul(g[1], a), fp_sub(0, t2), fp_sub(0, t3)],
-            fp_sqr(a),
+            [fp.fp_mul(g[0], a),fp.fp_mul(g[1], a),fp.fp_sub(0, t2),fp.fp_sub(0, t3)],
+           fp.fp_sqr(a),
         )
 
     else:
@@ -65,7 +71,7 @@ def reciprocal(f, flen, n):
         t = poly_mul_modxn(n - m, g, m, t[m:], n - m)                           # (f * g - a) * g mod x^n
 
         # Finally, the reciprocal is a*g - (f*g - a)*g mod x^n
-        t = [ fp_mul(a, g[i]) for i in range(0, m, 1) ] + [ fp_sub(0, t[i]) for i in range(0, n - m, 1) ]
+        t = [fp.fp_mul(a, g[i]) for i in range(0, m, 1) ] + [fp.fp_sub(0, t[i]) for i in range(0, n - m, 1) ]
         """
 
         # Basic idea but saving multiplication because of f*g - a is multiple of x^m
@@ -77,15 +83,17 @@ def reciprocal(f, flen, n):
         )  # (f * g - a) * g mod x^n
 
         # Finally, the reciprocal is a*g - (f*g - a)*g mod x^n
-        t = [fp_mul(a, g[i]) for i in range(0, m, 1)] + [
-            fp_sub(0, t[i]) for i in range(0, n - m, 1)
+        t = [fp.fp_mul(a, g[i]) for i in range(0, m, 1)] + [
+           fp.fp_sub(0, t[i]) for i in range(0, n - m, 1)
         ]
 
-        return t, fp_sqr(a)
+        return t,fp.fp_sqr(a)
 
 
-# Modular reduction in fp[x] with precomputation
 def poly_redc(h, hlen, f):
+    """
+    Modular reduction in fp[x] with precomputation
+    """
 
     flen = f['deg'] + 1
 
@@ -95,43 +103,43 @@ def poly_redc(h, hlen, f):
 
     elif flen == 2 and hlen == 2:
 
-        t0 = fp_mul(h[0], f['poly'][1])  # h0 * f1
-        t1 = fp_mul(h[1], f['poly'][0])  # h1 * f0
+        t0 =fp.fp_mul(h[0], f['poly'][1])  # h0 * f1
+        t1 =fp.fp_mul(h[1], f['poly'][0])  # h1 * f0
         return [fp_sub(t0, t1)]  # f1 * (h0 + h1*x) mod (f0 + f1*x)
 
     elif flen == 2 and hlen == 3:
 
-        f0_squared = fp_sqr(f['poly'][0])  # f0^2
-        f1_squared = fp_sqr(f['poly'][1])  # f1^2
-        t = fp_sub(f['poly'][0], f['poly'][1])  # f0 - f1
+        f0_squared =fp.fp_sqr(f['poly'][0])  # f0^2
+        f1_squared =fp.fp_sqr(f['poly'][1])  # f1^2
+        t =fp.fp_sub(f['poly'][0], f['poly'][1])  # f0 - f1
 
-        t = fp_sqr(t)  # (f0 - f1)^2
-        t = fp_sub(t, f0_squared)  # (f0 - f1)^2 - f0^2
-        t = fp_sub(t, f1_squared)  # (f0 - f1)^2 - f0^2 - f1^2 = -2*f0*f1
+        t =fp.fp_sqr(t)  # (f0 - f1)^2
+        t =fp.fp_sub(t, f0_squared)  # (f0 - f1)^2 - f0^2
+        t =fp.fp_sub(t, f1_squared)  # (f0 - f1)^2 - f0^2 - f1^2 = -2*f0*f1
 
-        f0_squared = fp_add(f0_squared, f0_squared)  # 2*(f0^2)
-        f1_squared = fp_add(f1_squared, f1_squared)  # 2*(f1^2)
+        f0_squared =fp.fp_add(f0_squared, f0_squared)  # 2*(f0^2)
+        f1_squared =fp.fp_add(f1_squared, f1_squared)  # 2*(f1^2)
 
-        t0 = fp_mul(f0_squared, h[2])  # [2*(f0^2)] * h2
-        t1 = fp_mul(f1_squared, h[0])  # [2*(f1^2)] * h0
-        t2 = fp_mul(t, h[1])  # [-2*f0*f1] * h1
+        t0 =fp.fp_mul(f0_squared, h[2])  # [2*(f0^2)] * h2
+        t1 =fp.fp_mul(f1_squared, h[0])  # [2*(f1^2)] * h0
+        t2 =fp.fp_mul(t, h[1])  # [-2*f0*f1] * h1
         return [
-            fp_add(t0, fp_add(t1, t2))
+           fp.fp_add(t0,fp.fp_add(t1, t2))
         ]  # [2 * (f1^2)] * (h0 + h1*x + h2*x^2) mod (f0 + f1*x)
 
     elif flen == 3 and hlen == 3:
 
-        f2h1 = fp_mul(f['poly'][2], h[1])
-        f2h0 = fp_mul(f['poly'][2], h[0])
-        f1h2 = fp_mul(f['poly'][1], h[2])
-        f0h2 = fp_mul(f['poly'][0], h[2])
-        return [fp_sub(f2h0, f0h2), fp_sub(f2h1, f1h2)]
-        """ 
+        f2h1 =fp.fp_mul(f['poly'][2], h[1])
+        f2h0 =fp.fp_mul(f['poly'][2], h[0])
+        f1h2 =fp.fp_mul(f['poly'][1], h[2])
+        f0h2 =fp.fp_mul(f['poly'][0], h[2])
+        return [fp.fp_sub(f2h0, f0h2),fp.fp_sub(f2h1, f1h2)]
+        """
     elif flen == hlen:
 
-        t0 = [ fp_mul(h[inner], f['poly'][fdeg - 1]) for inner in range(0, fdeg - 1, 1) ]
-        t1 = [ fp_mul(h[fdeg - 1], f['poly'][inner]) for inner in range(0, fdeg - 1, 1) ]
-        return [ fp_sub(t1[inner], t0[inner]) for inner in range(0, fdeg - 1, 1) ]
+        t0 = [fp.fp_mul(h[inner], f['poly'][fdeg - 1]) for inner in range(0, fdeg - 1, 1) ]
+        t1 = [fp.fp_mul(h[fdeg - 1], f['poly'][inner]) for inner in range(0, fdeg - 1, 1) ]
+        return [fp.fp_sub(t1[inner], t0[inner]) for inner in range(0, fdeg - 1, 1) ]
         """
     else:
 
@@ -156,13 +164,14 @@ def poly_redc(h, hlen, f):
 
         # a*h(x) - (q * f)(x) will gives a polynomial of degree (deg(f) - 1)
         return [
-            fp_sub(fp_mul(f['a'], h[i]), qf[i]) for i in range(0, flen - 1, 1)
+           fp.fp_sub(fp.fp_mul(f['a'], h[i]), qf[i]) for i in range(0, flen - 1, 1)
         ]
 
 
-# Reciprocal tree of a given product tree
 def reciprocal_tree(r, glen, ptree_f, n):
-
+    """
+    Reciprocal tree of a given product tree
+    """
     if n == 0:
         # Super special base case (nothing to do)
         return {
@@ -253,8 +262,10 @@ def reciprocal_tree(r, glen, ptree_f, n):
     }
 
 
-# Next function computes g(x) mod f_1(x), ..., g(x) mod f_n(x)
 def multieval_unscaled(g, glen, ptree_f, n):
+    """
+    Next function computes g(x) mod f_1(x), ..., g(x) mod f_n(x)
+    """
 
     if n == 0:
         return [[1]]
@@ -277,8 +288,10 @@ def multieval_unscaled(g, glen, ptree_f, n):
         return left + right
 
 
-# Next functions computes the scaled remainder tree
 def multieval_scaled(g, glen, f, flen, ptree_f, n):
+    """
+    Next functions computes the scaled remainder tree
+    """
 
     if n == 0:
         return [[1]]
