+++ sidh/fp.py
@@ -1,188 +1,152 @@
-from sidh.framework import *
+from functools import reduce
+from .constants import sop_data, parameters
+from .math import bitlength
 
-if (
-    setting.formulaes != 'tvelu'
-    and setting.formulaes != 'svelu'
-    and setting.formulaes != 'hvelu'
-) or (setting.algorithm != 'csidh' and setting.algorithm != 'bsidh'):
+class F_p(object):
+    def __init__(self, algorithm, prime):
+        # counters for field operations performed
+        self.fpadd = 0
+        self.fpsqr = 0
+        self.fpmul = 0
 
-    print("  ,-~~-.___.          ")
-    print(
-        " / |  '     \\        SYNTAX ERROR ..., run python3 %s -h for help"
-        % sys.argv[0]
-    )
-    print("(  )         0        ")
-    print(" \_/-, ,----'         ")
-    print("    ====           // ")
-    print("   /  \-'~;    /~~~(O)")
-    print("  /  __/~|   /       |")
-    print("=(  _____| (_________|")
-    exit(2)
+        if algorithm == 'csidh':
+            L = parameters['csidh'][prime]['L']
+            n = parameters['csidh'][prime]['n']
+            exponent_of_two = parameters['csidh'][prime]['exponent_of_two']
+            self.L = L
+            self.n = n
+            self.exponent_of_two = exponent_of_two
 
-bitlength = lambda x: len(bin(x)[2:])  # number of bits
-hamming_weight = lambda x: bin(x).count("1")
-# hamming weight: number of bits equal 1
+        else:
+            assert False, "bsidh not refactored yet"
 
-sign = lambda x: (1, -1)[x < 0]  # Sign of an integer
-isequal = {True: 1, False: 0}  # Simulating constant-time integer comparison
+            f = open(sop_data + prime)
 
-# --------------------------------------------------------------------------------------------------------------------------------
-try:
+            # The prime to be used
+            p = f.readline()
+            self.p = int(p, 16)
 
-    if setting.algorithm == 'csidh':
+            # List corresponding (p + 1)
+            Lp = f.readline()
+            Lp = [int(lp) for lp in Lp.split()]
+            # exponent_of_twop = Lp[0]
+            # Lp = Lp[1:]
+            Ep = f.readline()
+            Ep = [int(ep) for ep in Ep.split()]
+            assert len(Ep) == len(Lp)
+            np = len(Lp)
 
-        # List of Small odd primes, L := [l_0, ..., l_{n-1}]
-        f = open(sop_data + setting.prime)
-        L = f.read()
-        L = [int(l) for l in L.split()]
-        exponent_of_two = L[0]  #   Exponent of cofactor-2
-        L = list(L[1:])  #   Small Odd Primes l_i's
-        n = len(L)  #   Number of l_i's to be used
-        f.close()
+            # List corresponding (p - 1)
+            Lm = f.readline()
+            Lm = [int(lm) for lm in Lm.split()]
+            Em = f.readline()
+            Em = [int(em) for em in Em.split()]
+            assert len(Em) == len(Lm)
+            nm = len(Lm)
 
-    else:
+            f.close()
 
-        f = open(sop_data + setting.prime)
+        if algorithm == 'csidh':
+            p = (2 ** (exponent_of_two)) * reduce(
+                lambda x, y: (x * y), L
+            ) - 1  # p := 4 * l_0 * ... * l_n - 1
+            self.p = p
+            p_minus_one_halves = (p - 1) // 2  # (p - 1) / 2
+            validation_stop = sum([bitlength(l_i) for l_i in L]) / 2.0 + 2
+            self.p_minus_one_halves = p_minus_one_halves
+            self.validation_stop = validation_stop
+        else:
 
-        # The prime to be used
-        p = f.readline()
-        p = int(p, 16)
+            # pp = (2**exponent_of_twop) * reduce(lambda x,y : (x*y), [ Lp[i]**Ep[i] for i in range(0, np, 1)  ])
+            pp = reduce(
+                lambda x, y: (x * y), [Lp[i] ** Ep[i] for i in range(0, np, 1)]
+            )
+            pm = reduce(
+                lambda x, y: (x * y), [Lm[i] ** Em[i] for i in range(0, nm, 1)]
+            )
+            assert (p + 1) % pp == 0
+            assert (p - 1) % pm == 0
 
-        # List corresponding (p + 1)
-        Lp = f.readline()
-        Lp = [int(lp) for lp in Lp.split()]
-        # exponent_of_twop = Lp[0]
-        # Lp = Lp[1:]
-        Ep = f.readline()
-        Ep = [int(ep) for ep in Ep.split()]
-        assert len(Ep) == len(Lp)
-        np = len(Lp)
+            if p % 4 == 1:
+                print("// Case p = 1 mod 4 is not implemented yet!")
+                exit(-1)
 
-        # List corresponding (p - 1)
-        Lm = f.readline()
-        Lm = [int(lm) for lm in Lm.split()]
-        Em = f.readline()
-        Em = [int(em) for em in Em.split()]
-        assert len(Em) == len(Lm)
-        nm = len(Lm)
-
-        f.close()
-
-except IOError:
-
-    print("  ,-~~-.___.          ")
-    print(" / |  '     \\        SYNTAX ERROR ..., file not accessible")
-    print("(  )         0        ")
-    print(" \_/-, ,----'         ")
-    print("    ====           // ")
-    print("   /  \-'~;    /~~~(O)")
-    print("  /  __/~|   /       |")
-    print("=(  _____| (_________|")
-    exit(2)
-
-if setting.algorithm == 'csidh':
-
-    p = (2 ** (exponent_of_two)) * reduce(
-        lambda x, y: (x * y), L
-    ) - 1  # p := 4 * l_0 * ... * l_n - 1
-    p_minus_one_halves = (p - 1) // 2  # (p - 1) / 2
-
-    validation_stop = sum([bitlength(l_i) for l_i in L]) / 2.0 + 2
-else:
-
-    # pp = (2**exponent_of_twop) * reduce(lambda x,y : (x*y), [ Lp[i]**Ep[i] for i in range(0, np, 1)  ])
-    pp = reduce(
-        lambda x, y: (x * y), [Lp[i] ** Ep[i] for i in range(0, np, 1)]
-    )
-    pm = reduce(
-        lambda x, y: (x * y), [Lm[i] ** Em[i] for i in range(0, nm, 1)]
-    )
-    assert (p + 1) % pp == 0
-    assert (p - 1) % pm == 0
-
-    if p % 4 == 1:
-        print("// Case p = 1 mod 4 is not implemented yet!")
-        exit(-1)
-
-    p_minus_one_halves = (p - 1) // 2
-    p_minus_3_quarters = (p - 3) // 4
-
-# --------------------------------------------------------------------------------------------------------------------------------
-# Checking if p is composite
+            self.p_minus_one_halves = (p - 1) // 2
+            p_minus_3_quarters = (p - 3) // 4
 
 
-def _try_composite(a, d, n, s):
-    if pow(a, d, n) == 1:
-        return False
-    for i in range(s):
-        if pow(a, 2 ** i * d, n) == n - 1:
-            return False
-    return True  # n  is definitely composite
+    def set_zero_ops(self):
+
+        self.fpadd -= self.fpadd
+        self.fpsqr -= self.fpsqr
+        self.fpmul -= self.fpmul
 
 
-def is_prime(n):
-    """
-    Miller-Rabin primality test.
- 
-    A return value of False means n is certainly not prime. A return value of
-    True means n is very likely a prime.
-    """
-    if n != int(n):
-        return False
-    n = int(n)
-    # Miller-Rabin test for prime
-    if n == 0 or n == 1 or n == 4 or n == 6 or n == 8 or n == 9:
-        return False
+    def show_ops(label, a, b, flag):
 
-    if n == 2 or n == 3 or n == 5 or n == 7:
-        return True
-    s = 0
-    d = n - 1
-    while d % 2 == 0:
-        d >>= 1
-        s += 1
-    assert 2 ** s * d == n - 1
+        print("| %s: %7dM + %7dS + %7da" % (label, self.fpmul, self.fpsqr, self.fpadd), end="\t")
 
-    def trial_composite(a):
-        if pow(a, d, n) == 1:
-            return False
-        for i in range(s):
-            if pow(a, 2 ** i * d, n) == n - 1:
-                return False
-        return True
+        return None
 
-    bar = Bar('// Primality test on p', max=128)
-    for i in range(128):  # number of trials
-        a = random.randrange(2, n)
-        if trial_composite(a):
-            return False
-        bar.next()
-    bar.finish()
+    def get_ops(self):
+        return [self.fpmul, self.fpsqr, self.fpadd]
 
-    return True
+    # Modular inverse
+    def fp_inv(self, a):
+        g, x, y = xgcd(a, self.p)
+        # if g != 1:
+        # 	raise ValueError
+        return x % self.p
 
 
-if is_prime(p) == False:
-    print(
-        "[ERROR]\tThe integer number p := 4 * l_1, * ... * l_n - 1 is not prime where L := ",
-        L,
-    )
-    exit(3)
+    # Modular addition
+    def fp_add(self, a, b):
+        self.fpadd += 1
+        return (a + b) % self.p
 
-"""
-if( (sys.argv[0] != 'header.py') ):
-    print("/*")
-    print("The prime number to be used has the following form\n")
-    print("           %3d" % n)
-    print("        ~~~~~~~~~")
-    print("         |     | ")
-    print("p := 4 x |     | l_j   -   1, where each l_j is a small odd prime")
-    print("           j=1  ")
-    print("*/\n")
-"""
 
-# At this point, p is very likely a prime. Thus, we can continue
-# --------------------------------------------------------------------------------------------------------------------------------
+    # Modular substraction
+    def fp_sub(self, a, b):
+        self.fpadd += 1
+        return (a - b) % self.p
+
+
+    # Modular multiplication
+    def fp_mul(self, a, b):
+        self.fpmul += 1
+        return (a * b) % self.p
+
+
+    # Modular squaring
+    def fp_sqr(self, a):
+        self.fpsqr += 1
+        # print(fpsqr)
+        return (a ** 2) % self.p
+
+
+    # constant-time swap
+    def fp_cswap(self, x, y, b):
+
+        z = list([x, y])
+        z = list(z[:: (1 - 2 * b)])
+        return z[0], z[1]
+
+
+    # Modular exponentiation
+    def fp_exp(self, a, e):
+
+        bits_of_e = bitlength(e)
+        bits_of_e -= 1
+        tmp_a = a
+        # left-to-right method for computing a^e
+        for j in range(1, bits_of_e + 1):
+
+            tmp_a = self.fp_sqr(tmp_a)
+            if ((e >> (bits_of_e - j)) & 1) != 0:
+                tmp_a = self.fp_mul(tmp_a, a)
+
+        return tmp_a
+
 
 # Jacobi symbol used for checking if an integer has square-root in fp
 def jacobi(a, n):
@@ -223,107 +187,6 @@ def xgcd(aa, bb):
         lasty * (-1 if bb < 0 else 1),
     )
 
-
-# counters for field operations performed
-fpadd = 0
-fpsqr = 0
-fpmul = 0
-
-
-def set_zero_ops():
-
-    global fpadd
-    global fpsqr
-    global fpmul
-    fpadd -= fpadd
-    fpsqr -= fpsqr
-    fpmul -= fpmul
-
-
-def show_ops(label, a, b, flag):
-
-    global fpadd
-    global fpsqr
-    global fpmul
-
-    print("| %s: %7dM + %7dS + %7da" % (label, fpmul, fpsqr, fpadd), end="\t")
-
-    return None
-
-
-def get_ops():
-
-    global fpadd
-    global fpsqr
-    global fpmul
-
-    return [fpmul, fpsqr, fpadd]
-
-
-# Modular inverse
-def fp_inv(a):
-
-    g, x, y = xgcd(a, p)
-    # if g != 1:
-    # 	raise ValueError
-    return x % p
-
-
-# Modular addition
-def fp_add(a, b):
-
-    global fpadd
-    fpadd += 1
-    return (a + b) % p
-
-
-# Modular substraction
-def fp_sub(a, b):
-
-    global fpadd
-    fpadd += 1
-    return (a - b) % p
-
-
-# Modular multiplication
-def fp_mul(a, b):
-
-    global fpmul
-    fpmul += 1
-    return (a * b) % p
-
-
-# Modular squaring
-def fp_sqr(a):
-
-    global fpsqr
-    fpsqr += 1
-    # print(fpsqr)
-    return (a ** 2) % p
-
-
-# constant-time swap
-def fp_cswap(x, y, b):
-
-    z = list([x, y])
-    z = list(z[:: (1 - 2 * b)])
-    return z[0], z[1]
-
-
-# Modular exponentiation
-def fp_exp(a, e):
-
-    bits_of_e = bitlength(e)
-    bits_of_e -= 1
-    tmp_a = a
-    # left-to-right method for computing a^e
-    for j in range(1, bits_of_e + 1):
-
-        tmp_a = fp_sqr(tmp_a)
-        if ((e >> (bits_of_e - j)) & 1) != 0:
-            tmp_a = fp_mul(tmp_a, a)
-
-    return tmp_a
 
 
 # --------------------------------------------------------------------------------------------------------------------------------
